backend/requirements.txt
# Core framework
strands-agents>=1.0.0
strands-agents-tools>=0.2.0

# AWS and configuration
boto3>=1.26.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
aiofiles>=23.2.1

# Visualization and geo
folium>=0.14.0
geopy>=2.3.0

# Data modeling
pydantic>=2.0.0

# Optional security/PII protection (commenting out problematic packages)
# llm-guard>=0.3.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Add additional packages with specific versions if needed